buildsense_back/core/serializers.py
# buildsense_back/core/serializers.py
from rest_framework import serializers
from .models import Company, Project

class CompanySerializer(serializers.ModelSerializer):
    class Meta:
        model = Company
        fields = ['id', 'name']

class ProjectSerializer(serializers.ModelSerializer):
    class Meta:
        model = Project
        fields = ['id', 'name', 'company']